backend/pyproject.toml
[tool.poetry]
name = "agreement-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "../backend" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
g4f = "0.4.9.1"
langchain = "^0.3.19"
langgraph = "^0.2.74"
pypandoc = "^1.15"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
websockets = "^15.0"
openai = "^1.64.0"
prisma = "^0.15.0"
email-validator = "^2.2.0"
pillow = "^11.1.0"
langchain-openai = "^0.3.7"
python-jose = "^3.4.0"
curl-cffi = "^0.9.0"
langchain-core = "^0.3.40"
pymupdf = "^1.25.3"
face-recognition = "^1.3.0"
numpy = "^2.2.3"
face-recognition-models = {git = "https://github.com/ageitgey/face_recognition_models.git"}
reportlab = "^4.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
